# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Teamwork API
  description: An API for an internal social network for employees of an organization.
  version: 1.0.0
servers:
  - url: https://team-worker.herokuapp.com/api/v1
    description: Main (production) server
  - url: http://localhost:3000/api/v1
    description: Internal development server for local testing
...
paths:
...
  /articles:
    post:
      tags:
        - article
      operationId: createArticle
      summary: Create a new article
      description: Create a new article
      requestBody:
        description: Create a new article
        content:
          application/json:
            schema: 
              type: object
              required:
                - title
                - article
              properties:
                title:
                  type: string
                  description: Title of article
                  example: How to become a 10x software developer
                article:
                  type: string
                  description: Article content
                  example: /You are expected to create a set of ​ API endpoints​ listed under the ​ API Endpoints Specification\
                      section and ensure you persist your data using a PostgreSQL database.\
                      You are to write SQL queries that will help you write to and read from your database. The
                      endpoints are to be secured with JSON Web Token (JWT)
      responses:
        '201':
          description: A successful operation
          content: 
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: successful
                    example: success
                  data:
                    type: object
                    description: Data for new article created
                    properties:
                      message:
                        type: string
                        description: Message for article created
                        example: Article successfully posted
                      articleId:
                        type: integer
                        description: Article id
                        example: 10
                      createdOn: 
                        type: string
                        description: Created date of a new article
                        pattern: /([0-9]{4})-(?:[0-9]{2})-([0-9]{2})/
                        example: "2019-05-17"
                      title: 
                        type: string
                        description: Title of article
                        example: How to become a 10x software developer
      security:
        - access_token: []
  /articles/{articleId}:
    patch:
      tags:
        - article
      operationId: updateArticle
      summary: Edit article
      description: Edit article with an article id
      requestBody:
        description: Edit a new article by article userId
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Title of the article
                  example: How to be a 10x Frontend developer
                article:
                  type: string
                  description: Content of the article
                  example: You need to design and build 10x spectacular projects 
      responses:
        '201':
          description: A Successful Operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: A successful content update
                    example: success
                  data:
                    type: object
                    description: Article returned on a successful article update
                    properties:
                      message:
                        type: string
                        description: Message to indicate successful update
                        example: Article successfully updated 
                      title:
                        type: string
                        description: Title of the article
                        example: How to be a 10x Frontend developer
                      article:
                        type: string
                        description: Content of the article
                        example: You need to design and build 10x spectacular projects
      security:
        - access_token: [] 
    delete:
      tags:
        - article
      operationId: deleteArticle
      summary: Delete article
      description: Delete article with an article id
      responses:
        '204':
          description: A Successful Operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: success
                    example: success
                  data:
                    type: object
                    description: Article deleted successfuly
                    properties:
                      message:
                        type: string
                        description: Message to indicate successful delete
                        example: Article successfully deleted 
      security:
        - access_token: []
    get:
      tags:
        - article
      operationId: getArticle
      summary: Get an article
      description: Get an article with an article id
      responses:
        '200':
          description: A successfully Operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: success
                    example: success
                  data:
                    type: object
                    description: Article deleted successfuly
                    properties:
                      id:
                        type: integer
                        description: Article id
                        example: 10
                      createdOn: 
                        type: string
                        description: Created date of a new article
                        pattern: /([0-9]{4})-(?:[0-9]{2})-([0-9]{2})/
                        example: "2019-05-17"
                      articleTitle: 
                        type: string
                        description: Title of article
                        example: How to become a 10x software developer
                      article: 
                        type: string
                        description: Article comment was created for
                        example: /You are expected to create a set of ​ API endpoints​ listed under the ​ API Endpoints Specification\
                          section and ensure you persist your data using a PostgreSQL database.\
                          You are to write SQL queries that will help you write to and read from your database. The
                          endpoints are to be secured with JSON Web Token (JWT)
                      comments:
                        type: array
                        items:
                          properties:
                            commentId:
                              type: integer
                              description: Comment Id
                              example: 10
                            comment:
                              type: string
                              description: New comment to be created
                              example: This post articulates all there is to be a 10x software developer
                            authorId:
                              type: integer
                              description: Comment Id
                              example: 10
      security:
          - access_token: []
  /articles/{articleId}/comment:
    post:
      tags:
        - article
      operationId: Create a comment for an article
      summary: Create a comment for an article
      description: Create a comment for an article
      requestBody:
        description: Create a new comment for a post
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                  description: New comment to be created
                  example: This post articulates all there is to be a 10x software developer
      responses:
        '201':
          description: Successful creation of comment for article
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: successful creation of comment
                    example: success
                  data:
                    type: object
                    description: Data for new comment created on article
                    properties:
                      message:
                        type: string
                        description: Message for comment created on article.
                        example: Comment successfully created
                      createdOn: 
                        type: string
                        description: Created date of a new article
                        pattern: /([0-9]{4})-(?:[0-9]{2})-([0-9]{2})/
                        example: "2019-05-17"
                      articleTitle: 
                        type: string
                        description: Title of article
                        example: How to become a 10x software developer
                      article: 
                        type: string
                        description: Article comment was created for
                        example: /You are expected to create a set of ​ API endpoints​ listed under the ​ API Endpoints Specification\
                          section and ensure you persist your data using a PostgreSQL database.\
                          You are to write SQL queries that will help you write to and read from your database. The
                          endpoints are to be secured with JSON Web Token (JWT)
                      comment:
                        type: string
                        description: New comment to be created
                        example: This post articulates all there is to be a 10x software developer
      security:
        - access_token: []
  /articles/{articleId}/likes:
    post:
      tags:
        - article
      operationId: Record a like for an article
      summary: Record a like for an article
      description: Record a like for an article
      parameters:
      - in : path
        name : articleId
        schema:
          type: integer
        required: true
        description : Numeric ID of an article to get
      requestBody:
        description: Create a new like for an article
        content:
          application/json:
            schema:
              type: object
              properties:
                userId: 
                  type: integer
                  description: user id
                  example: 10
      responses:
        '201':
          description: Successful recording a like for an article
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: successful liking an article
                    example: success
                  data:
                    type: object
                    description: Data for new like created on an article
                    properties:
                      message:
                        type: string
                        description: Message for like created on an article
                        example: Article  successfully liked
                      userId: 
                        type: integer
                        description: user id
                        example: 10
                      articleId: 
                        type: integer
                        description: article id
                        example: 10
      security:
        - access_token: []
...